Refactor dataset and datastore connection slices to inherit from baseAPI
This allows us to invalidate the cache on datasets from within the datastore connection slice.
